Candidate model, new api calls & tests reorganized
