+++ HiringApp/HiringAppCore/HiringAppCore.podspec
@@ -18,6 +18,6 @@ Pod::Spec.new do |spec|
   }
   spec.dependency 'BSWFoundation', '~> 1.0.1'
   spec.dependency 'RealmSwift'
-
+  spec.dependency 'KeychainSwift'
 
 end
