+++ HiringApp/HiringAppCore/Providers/APIProvider.swift
@@ -10,21 +10,12 @@ import Foundation
 import BSWFoundation
 import Deferred
 import Decodable
+import KeychainSwift
 
 enum APIError: Error {
     case badStatusCode
     case errorWhileParsing
-}
-
-struct DevelopmentEnvironment: Environment {
-    var basePath: String {
-        return "http://private-60454-visualtestapp.apiary-mock.com"
-    }
-}
-
-struct TechologiesEndpoint: Endpoint {
-    /// The path for the request
-    var path: String { return "/technologies" }
+    case responseUnexpected
 }
 
 protocol DroskyType {
@@ -73,6 +64,38 @@ class APIProvider: APIProviderType {
         return modelTask
     }
     
-    
+    func performLogin() -> Task<String> {
+        
+        //perform a request
+        let droskyTask = drosky.performRequest(forEndpoint: AuthenticateEndpoint())
+        
+        //check the status code of the response
+        let statusCodeTask = droskyTask.andThen(upon: .global()) { (droskyResponse) -> Task<Data> in
+            //Check if status code is in range od 200s
+            guard droskyResponse.statusCode >= 200 && droskyResponse.statusCode < 300 else {
+                return Task(failure: APIError.badStatusCode)
+            }
+            
+            return Task(success: droskyResponse.data)
+        }
+        
+        //parse the data of the response and return a model
+        let modelTask = statusCodeTask.andThen(upon: .global()) { (data) -> Task<String> in
+            do {
+                let json = try JSONSerialization.jsonObject(with: data, options: []) as? Dictionary<String, Any>
+                guard let token = json?["authToken"] as? String else {
+                    return Task(failure: APIError.errorWhileParsing)
+                }
+                
+                let keychain = KeychainSwift()
+                keychain.set(token, forKey: "userToken")
+                print(keychain.get("userToken"))
+                return Task(success: token)
+            } catch {
+                return Task(failure: APIError.errorWhileParsing)
+            }
+        }
+        return modelTask
+    }
     
 }
